♻️ refactor[#3] : 패키지 오타 application 수정
