+++ src/main/java/com/fouribnb/notification/application/dto/responseDto/NotificationInternalResponse.java
@@ -1,4 +1,4 @@
-package com.fouribnb.notification.applicatoin.dto.responseDto;
+package com.fouribnb.notification.application.dto.responseDto;
 
 import com.fouribnb.notification.domain.enums.Type;
 import java.util.UUID;
