+++ src/main/java/com/fouribnb/notification/application/service/NotificationService.java
@@ -1,0 +1,9 @@
+package com.fouribnb.notification.application.service;
+
+import com.fouribnb.notification.application.dto.requestDto.CreateNotificationInternalRequest;
+import com.fouribnb.notification.application.dto.responseDto.NotificationInternalResponse;
+
+public interface NotificationService {
+
+    NotificationInternalResponse createNotification(CreateNotificationInternalRequest request);
+}
