+++ src/main/java/com/fouribnb/notification/application/service/NotificationServiceImpl.java
@@ -1,8 +1,8 @@
-package com.fouribnb.notification.applicatoin.service;
+package com.fouribnb.notification.application.service;
 
-import com.fouribnb.notification.applicatoin.dto.requestDto.CreateNotificationInternalRequest;
-import com.fouribnb.notification.applicatoin.dto.responseDto.NotificationInternalResponse;
-import com.fouribnb.notification.applicatoin.mapper.NotificationMapper;
+import com.fouribnb.notification.application.dto.requestDto.CreateNotificationInternalRequest;
+import com.fouribnb.notification.application.dto.responseDto.NotificationInternalResponse;
+import com.fouribnb.notification.application.mapper.NotificationMapper;
 import com.fouribnb.notification.domain.entity.Notification;
 import com.fouribnb.notification.domain.repository.NotificationRepository;
 import lombok.RequiredArgsConstructor;
