--- src/main/java/com/fouribnb/notification/applicatoin/service/NotificationService.java
@@ -1,9 +1,0 @@
-package com.fouribnb.notification.applicatoin.service;
-
-import com.fouribnb.notification.applicatoin.dto.requestDto.CreateNotificationInternalRequest;
-import com.fouribnb.notification.applicatoin.dto.responseDto.NotificationInternalResponse;
-
-public interface NotificationService {
-
-    NotificationInternalResponse createNotification(CreateNotificationInternalRequest request);
-}
